requirements.txt
-i https://pypi.org/simple
amqp==2.5.0
babel==2.7.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
click==7.0
dnspython==1.16.0
eventlet==0.25.0
flask==1.0.3
flower==0.9.3
future==0.17.1
futures==3.1.1
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.1
kombu==4.6.3
markupsafe==1.1.1
mongoengine==0.18.2
monotonic==1.5
numpy==1.16.4
pillow==6.0.0
pymongo==3.8.0
pytz==2019.1
redis==3.2.1
requests==2.22.0
six==1.12.0
tornado==5.1.1
urllib3==1.25.3
vine==1.3.0
werkzeug==0.15.4
